pyproject.toml
[project]
name = "ynym-portal-backend"
version = "0.1.0"
description = "FastAPI backend system for ynym portal"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]==0.135.*",
    "sqlalchemy>=2.0.48",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "psycopg[binary]>=3.3.3",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "greenlet>=3.2.4",
]

[dependency-groups]
dev = [
    "pytest",
    "httpx",
    "ruff",
    "mypy",
    "mkdocs",
    "mkdocs-material",
    "pytest-mock",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--strict-markers"
markers = [
    "unit: Mark test as unit test",
    "integration: Mark test as integration test",
]
